requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
sqlalchemy==2.0.36
apscheduler==3.10.4
feedparser==6.0.11
httpx==0.28.1
beautifulsoup4==4.12.3
openai==1.51.2
streamlit==1.39.0
plotly==5.24.1
pandas==2.2.3
yfinance==0.2.54
pytest==8.3.3
google-genai==1.35.0
chromadb==0.5.23
